+++ sphinx/application/common/wrappers/wrapper-podcast/src/main/java/chat/sphinx/wrapper_podcast/PodcastModel.kt
@@ -1,0 +1,6 @@
+package chat.sphinx.wrapper_podcast
+
+data class PodcastModel(
+    val type: String,
+    val suggested: Double,
+)