+++ sphinx/application/network/concepts/queries/concept-network-query-chat/src/main/java/chat/sphinx/concept_network_query_chat/model/PodcastEpisodeDto.kt
@@ -1,5 +1,6 @@
 package chat.sphinx.concept_network_query_chat.model
 
+import chat.sphinx.wrapper_podcast.PodcastEpisode
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
@@ -10,5 +11,8 @@ data class PodcastEpisodeDto(
     val image: String,
     val link: String,
     val enclosureUrl: String,
-) {
+)
+
+fun PodcastEpisodeDto.toPodcastEpisode(): PodcastEpisode {
+    return PodcastEpisode(this.id, this.title ,this.description, this.image, this.link, this.enclosureUrl)
 }