+++ sphinx/screens/chats/chat-common/chat-common/src/main/java/chat/sphinx/chat_common/ui/viewstate/header/ChatHeaderViewState.kt
@@ -12,7 +12,6 @@ sealed class ChatHeaderViewState: ViewState<ChatHeaderViewState>() {
         val chatHeaderName: String,
         val showLock: Boolean,
         val showExitTribe: Boolean,
-        val contributions: Sat?,
         val isMuted: ChatMuted?,
     ): ChatHeaderViewState()
 }