+++ sphinx/screens/chats/chat-tribe/chat-tribe/src/main/java/chat/sphinx/chat_tribe/navigation/TribeChatNavigator.kt
@@ -2,7 +2,7 @@ package chat.sphinx.chat_tribe.navigation
 
 import androidx.navigation.NavController
 import chat.sphinx.chat_common.navigation.ChatNavigator
-import chat.sphinx.podcast_player.objects.Podcast
+import chat.sphinx.podcast_player.objects.ParcelablePodcast
 import chat.sphinx.wrapper_common.dashboard.ChatId
 import io.matthewnelson.concept_navigation.BaseNavigationDriver
 
@@ -10,5 +10,5 @@ abstract class TribeChatNavigator(
     navigationDriver: BaseNavigationDriver<NavController>
 ): ChatNavigator(navigationDriver)
 {
-    abstract suspend fun toPodcastPlayerScreen(chatId: ChatId, podcast: Podcast)
+    abstract suspend fun toPodcastPlayerScreen(chatId: ChatId, podcast: ParcelablePodcast)
 }