+++ sphinx/screens/chats/chat-tribe/chat-tribe/src/main/java/chat/sphinx/chat_tribe/ui/PodcastViewState.kt
@@ -1,6 +1,7 @@
 package chat.sphinx.chat_tribe.ui
 
-import chat.sphinx.podcast_player.objects.Podcast
+import chat.sphinx.podcast_player.objects.ParcelablePodcast
+import chat.sphinx.wrapper_podcast.Podcast
 import io.matthewnelson.concept_views.viewstate.ViewState
 
 sealed class PodcastViewState: ViewState<PodcastViewState>() {